refactor: move index, supersede, status, describe verbs to cli commands modules
