+++ cli/src/commands/mod.rs
@@ -1,5 +1,9 @@
 //! One module per CLI subcommand wrapper, mirroring `living_docs_core::commands` (issue 0028).
 
 pub(crate) mod brief;
+pub(crate) mod describe;
+pub(crate) mod index;
 pub(crate) mod new;
 pub(crate) mod next;
+pub(crate) mod status;
+pub(crate) mod supersede;
